cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sqlazo"
version = "0.1.0"
description = "Execute SQL queries from files with connection headers"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "angoru"}
]
dependencies = [
    "mysql-connector-python>=8.0",
    "psycopg2-binary>=2.9",
    "pymongo>=4.0",
    "redis>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
sqlazo = "sqlazo.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
